# internal/handler/openapi.yaml
openapi: "3.0.3"
info:
  title: CronControl API
  description: |
    Control plane for scheduled and event-driven operational workloads.
    API-first, agentic-ready. Uses canonical terminology: workspace, process, run, queue, job, worker.
  version: "1.0.0"
servers:
  - url: /api/v1
security:
  - ApiKeyAuth: []
  - CookieAuth: []
paths:
...
  /register:
    post:
      operationId: register
      summary: Register a new workspace
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Workspace created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/RateLimited"
  /register/verify:
    post:
      operationId: verifyEmail
      summary: Verify email address
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
      responses:
        "200":
          description: Email verified
        "400":
          $ref: "#/components/responses/BadRequest"

  # =========================================================================
  # HEARTBEAT (no auth — implicit via run ID)
  # =========================================================================
  /heartbeat:
    post:
      operationId: sendHeartbeat
      summary: Report heartbeat and progress (non-HTTP methods only)
      tags: [Heartbeat]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HeartbeatRequest"
      responses:
        "200":
          description: Heartbeat recorded
        "404":
          $ref: "#/components/responses/NotFound"

  # =========================================================================
  # WORKSPACE
  # =========================================================================
  /workspace:
    get:
      operationId: getWorkspace
      summary: Get current workspace info
      tags: [Workspace]
      responses:
        "200":
          description: Workspace details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SingleResponse_Workspace"
    put:
      operationId: updateWorkspace
      summary: Update workspace settings (admin)
      tags: [Workspace]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkspaceRequest"
      responses:
        "200":
          description: Workspace updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SingleResponse_Workspace"
...
  /processes:
    get:
      operationId: listProcesses
      summary: List processes
      tags: [Processes]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
        - name: schedule_type
          in: query
          schema:
            type: string
            enum: [cron, fixed_delay, on_demand]
        - name: enabled
          in: query
          schema:
            type: boolean
        - name: tag
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Process list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListResponse_Process"
    post:
      operationId: createProcess
      summary: Create a process (admin)
      tags: [Processes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProcessRequest"
      responses:
        "201":
          description: Process created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SingleResponse_Process"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /processes/{id}:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    get:
      operationId: getProcess
      summary: Get process details
      tags: [Processes]
      responses:
        "200":
          description: Process details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SingleResponse_Process"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateProcess
      summary: Update a process (admin)
      tags: [Processes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProcessRequest"
      responses:
        "200":
          description: Process updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SingleResponse_Process"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteProcess
      summary: Delete a process (admin)
      tags: [Processes]
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /processes/{id}/trigger:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    post:
      operationId: triggerProcess
      summary: Manually trigger a process (operator)
      tags: [Processes]
      responses:
        "201":
          description: Run created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SingleResponse_Run"
        "404":
          $ref: "#/components/responses/NotFound"
  /processes/{id}/pause:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    post:
      operationId: pauseProcess
      summary: Pause scheduling (operator)
      tags: [Processes]
      parameters:
        - name: cancel_pending
          in: query
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Paused
        "404":
          $ref: "#/components/responses/NotFound"
  /processes/{id}/resume:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    post:
      operationId: resumeProcess
      summary: Resume scheduling (operator)
      tags: [Processes]
      responses:
        "200":
          description: Resumed
        "404":
          $ref: "#/components/responses/NotFound"

  # =========================================================================
  # RUNS
  # =========================================================================
  /runs:
    get:
      operationId: listRuns
      summary: List runs
      tags: [Runs]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
        - name: process_id
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
        - name: origin
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Run list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListResponse_Run"
    post:
      operationId: createRun
      summary: Schedule a one-time run (operator)
      tags: [Runs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRunRequest"
      responses:
        "201":
          description: Run scheduled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SingleResponse_Run"
  /runs/{id}:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    get:
      operationId: getRun
      summary: Get run details with attempts
      tags: [Runs]
      responses:
        "200":
          description: Run details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SingleResponse_RunDetail"
        "404":
          $ref: "#/components/responses/NotFound"
  /runs/{id}/kill:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    post:
      operationId: killRun
      summary: Request stop for an active execution (operator)
      tags: [Runs]
      responses:
        "200":
          description: Kill requested. The run may stay in `kill_requested` until the executor confirms remote stop and moves it to `killed`.
        "404":
          $ref: "#/components/responses/NotFound"
  /runs/{id}/cancel:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    post:
      operationId: cancelRun
      summary: Cancel a run or request stop if already active (operator)
      tags: [Runs]
      responses:
        "200":
          description: If the run has not started, it becomes `cancelled`. If already active, the API transitions it to `kill_requested` and the terminal state becomes `killed` only after confirmation.
        "404":
          $ref: "#/components/responses/NotFound"
  /runs/{id}/replay:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    post:
      operationId: replayRun
      summary: Replay a terminal run (operator)
      tags: [Runs]
      responses:
        "201":
          description: Replay run created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SingleResponse_Run"
        "404":
          $ref: "#/components/responses/NotFound"
  /runs/{id}/output:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    get:
      operationId: getRunOutput
      summary: Get run stdout/stderr
      tags: [Runs]
      parameters:
        - name: stream
          in: query
          schema:
            type: string
            enum: [stdout, stderr]
      responses:
        "200":
          description: Output
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunOutputResponse"
        "404":
          $ref: "#/components/responses/NotFound"

  # =========================================================================
  # QUEUES
  # =========================================================================
  /queues:
    get:
      operationId: listQueues
      summary: List queues
      tags: [Queues]
      responses:
        "200":
          description: Queue list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListResponse_Queue"
    post:
      operationId: createQueue
      summary: Create a queue (admin)
      tags: [Queues]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQueueRequest"
      responses:
        "201":
          description: Queue created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SingleResponse_Queue"
  /queues/{id}:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    get:
      operationId: getQueue
      summary: Get queue details with stats
      tags: [Queues]
      responses:
        "200":
          description: Queue details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SingleResponse_QueueDetail"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateQueue
      summary: Update queue (admin)
      tags: [Queues]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQueueRequest"
      responses:
        "200":
          description: Queue updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SingleResponse_Queue"
    delete:
      operationId: deleteQueue
      summary: Delete queue (admin)
      tags: [Queues]
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"

  # =========================================================================
  # JOBS
  # =========================================================================
...
  /jobs/{id}:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    get:
      operationId: getJob
      summary: Get job with attempt history
      tags: [Jobs]
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SingleResponse_JobDetail"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    CookieAuth:
      type: apiKey
      in: cookie
      name: session
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
...
    RegisterRequest:
      type: object
      required: [email, name]
      properties:
        email:
          type: string
          format: email
        name:
          type: string
        password:
          type: string
          minLength: 12
    RegisterResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            workspace:
              type: object
              properties:
                id:
                  type: string
                slug:
                  type: string
            user:
              type: object
              properties:
                id:
                  type: string
                email:
                  type: string
                role:
                  type: string
            api_key:
              type: string
              description: Shown once. Save immediately.
            email_verified:
              type: boolean
            hint:
              type: string
...